Split test_meta_adaptation.py into focused test modules (Commit 2)
Finding: the 3912-line monolithic test file had 19 classes spanning signals,
detection, builders/e2e, and verdict — no grouping by concern; also
_LAM_NONTRIVIAL_TOL was imported but unused in builders.py (leftover from
an earlier draft); inline imports in TestSharedEpsilonDA referenced the old
module path; LowRankInverseMassMatrix and _make_engine were caught as missing
only when running tests (scan missed non-meta_adaptation imports).
Fix: split into 4 focused files + shared _meta_fixtures.py; removed unused
_LAM_NONTRIVIAL_TOL from builders.py; all 100 meta tests pass 100/100.
