+++ blackjax/adaptation/meta/builders.py
@@ -36,7 +36,6 @@ from blackjax.adaptation.meta._calibration import (
     _DETECTION_BRANCH_NONE,
     _DETECTION_BRANCH_POOLED_WITHIN,
     _GAIN_THRESHOLD,
-    _LAM_NONTRIVIAL_TOL,
     _MAX_RANK_CAP,
     _MC_COLLINEARITY_TOL,
     _MC_MIN_CHAINS,
@@ -46,8 +45,8 @@ from blackjax.adaptation.meta._calibration import (
     _R2_DEFERRED,
     _R2_PROJECTED,
     _R_MIN,
-    _S_MIN,
     _S_GAP_STABILITY_TOL,
+    _S_MIN,
     _STEP_SIZE_READAPT_BUFFER,
     _W_BRANCH_R1_TOL,
     _mc_detection_edge,
